Added some food!
For now, this can be spawned in with F, but we need a button to spawn
the food in later (as well as a food picker, to determine which food we
want to spawn in).

As long as the egg is hungry, and there is food in the world, egg will
bounce towards food.

Once egg is close enough to food, it will enter an Eat state, whereupon
it eats the food, and food is removed from world.

Part of https://github.com/lucymhdavies/egg-game/issues/6